pyproject.toml
[tool.poetry]
name = "learning-platform"
version = "0.1.0"
description = ""
authors = [
            "SergeyPyt <virvich.serg.yandex.by>",
            "olchrkk <olchrkk@gmail.com>",
            "nikitapt63 <nikitapt63@gmail.com>",
          ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
pre-commit = "^3.4.0"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
psycopg2 = "^2.9.8"
pillow = "^10.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
